+++ global-client/src/components/ui/HubIcon.vue
@@ -1,6 +1,11 @@
 <template>
 	<div class="block group text-center mb-2 cursor-pointer relative" :title="hub ? hub.hubId : null">
-		<Badge v-if="hub && hub.unreadMessages > 0" class="sm:ml-6">{{ hub.unreadMessages }}</Badge>
+		<div v-if="hub && hub.unreadMessages > 0">
+			<span class="absolute flex h-3 w-3 left-8">
+				<span class="animate-ping absolute inline-flex h-full w-full rounded-full bg-red opacity-75"></span>
+				<span class="absolute inline-flex rounded-full h-3 w-3 bg-notification"></span>
+			</span>
+		</div>
 		<Icon v-if="pinnable" type="plus" class="text-green absolute right-0" @click.prevent="pin"></Icon>
 		<Icon v-if="pinned" type="remove" class="text-red absolute right-0 hidden group-hover:block" @click.prevent="remove"></Icon>
 		<Icon :type="type" :size="size" class="text-white mx-auto"></Icon>