+++ hub-client/src/components/elements/UnreadMessageBadge.vue
@@ -1,0 +1,7 @@
+<template>
+	<div class="relative">
+		<div class="bg-red text-white absolute rounded-full text-center min-w-5 min-h-5 p-0 text-sm">
+			<span class="inline-block"><slot></slot></span>
+		</div>
+	</div>
+</template>