+++ hub-client/src/components/rooms/RoomList.vue
@@ -5,9 +5,17 @@
 		<template v-if="rooms.hasRooms">
 			<template v-for="room in rooms.sortedRoomsArray" :key="room.roomId">
 				<div v-if="showRoom(room)" :key="room.roomId" class="group" @click="toggleMenu.toggleGlobalMenu()">
+					<template v-if="settings.isFeatureEnabled(featureFlagType.notifications)">
+						<UnreadMessageBadge v-if="room.getRoomUnreadNotificationCount(NotificationCountType.Total) > 0" class="ml-2 float-right my-2 mr-12">{{
+							room.getRoomUnreadNotificationCount(NotificationCountType.Total)
+						}}</UnreadMessageBadge>
+						<UnreadMentionBadge v-if="room.getRoomUnreadNotificationCount(NotificationCountType.Highlight) > 0" class="ml-2 float-right my-2 mr-4">{{
+							room.getRoomUnreadNotificationCount(NotificationCountType.Highlight)
+						}}</UnreadMentionBadge>
+					</template>
 					<Icon type="unlink" class="cursor-pointer invisible hover:text-red ml-2 float-right my-2 mr-8 group-hover:visible" @click="leaveRoom(room.roomId)"></Icon>
+
 					<router-link :to="{ name: 'room', params: { id: room.roomId } }" v-slot="{ isActive }">
-						<Badge v-if="room.numUnreadMessages > 0" class="-ml-1 -mt-1">{{ room.numUnreadMessages }}</Badge>
 						<MenuItem :roomInfo="room" :icon="roomIcon(room)" :active="isActive">
 							<PrivateRoomName v-if="room.isPrivateRoom()" :members="room.getOtherJoinedAndInvitedMembers()"></PrivateRoomName>
 							<span v-else>
@@ -29,6 +37,9 @@
 	import { RoomType } from '@/store/rooms';
 	import { usePlugins, PluginProperties } from '@/store/plugins';
 	import { useToggleMenu } from '@/store/toggleGlobalMenu';
+	import { NotificationCountType } from 'matrix-js-sdk';
+	import { useSettings, featureFlagType } from '@/store/store';
+	const settings = useSettings();
 
 	const { t } = useI18n();
 	const router = useRouter();